Add MongoDB utility with associated test and documentation
This commit includes the addition of a general-purpose utility for using MongoDB's Go driver, termed as 'Mongoutil'. It provides clearer, more readable code for your DB operations. Besides the full explanation in the README.md, this commit also includes the first set of tests to ensure the functionality of the utility and an MIT License file.